Move BeanMapper functionality to PojoMapper and flip the hierarchy, same with BeanPropertyArguments
