+++ core/src/main/java/org/jdbi/v3/core/argument/BeanPropertyArguments.java
@@ -13,13 +13,8 @@
  */
 package org.jdbi.v3.core.argument;
 
-import java.util.Optional;
-import org.jdbi.v3.core.argument.internal.MethodReturnValueNamedArgumentFinder;
-import org.jdbi.v3.core.argument.internal.TypedValue;
-import org.jdbi.v3.core.statement.StatementContext;
+import org.jdbi.v3.core.argument.internal.PojoPropertyArguments;
 import org.jdbi.v3.core.mapper.reflect.internal.BeanPropertiesFactory;
-import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties;
-import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties.PojoProperty;
 
 /**
  * Inspect a {@link java.beans} style object and bind parameters
@@ -28,46 +23,17 @@ import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties.PojoProperty;
  * @deprecated this should never have been public API
  */
 @Deprecated
-public class BeanPropertyArguments extends MethodReturnValueNamedArgumentFinder {
-    private final PojoProperties<?> properties;
-
+public class BeanPropertyArguments extends PojoPropertyArguments {
     /**
      * @param prefix an optional prefix (we insert a '.' as a separator)
      * @param bean the bean to inspect and bind
      */
     public BeanPropertyArguments(String prefix, Object bean) {
-        super(prefix, bean);
-        properties = BeanPropertiesFactory.propertiesFor(obj.getClass());
-    }
-
-    /**
-     * @param prefix an optional prefix (we insert a '.' as a separator)
-     * @param bean the bean to inspect and bind
-     */
-    protected BeanPropertyArguments(String prefix, Object bean, PojoProperties<?> properties) {
-        super(prefix, bean);
-        this.properties = properties;
-    }
-
-    @Override
-    protected Optional<TypedValue> getValue(String name, StatementContext ctx) {
-        @SuppressWarnings("unchecked")
-        PojoProperty<Object> property = (PojoProperty<Object>) properties.getProperties().get(name);
-
-        if (property == null) {
-            return Optional.empty();
-        }
-
-        return Optional.of(new TypedValue(property.getQualifiedType(), property.get(obj)));
+        super(prefix, bean, BeanPropertiesFactory.propertiesFor(bean.getClass()), null);
     }
 
     @Override
     protected NamedArgumentFinder getNestedArgumentFinder(Object o) {
         return new BeanPropertyArguments(null, o);
-    }
-
-    @Override
-    public String toString() {
-        return "{lazy bean property arguments \"" + obj + "\"}";
     }
 }
