+++ core/src/main/java/org/jdbi/v3/core/mapper/reflect/BeanMapper.java
@@ -13,29 +13,14 @@
  */
 package org.jdbi.v3.core.mapper.reflect;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-
 import org.jdbi.v3.core.generic.GenericTypes;
-import org.jdbi.v3.core.mapper.ColumnMapper;
-import org.jdbi.v3.core.mapper.Nested;
-import org.jdbi.v3.core.mapper.NoSuchMapperException;
 import org.jdbi.v3.core.mapper.RowMapper;
 import org.jdbi.v3.core.mapper.RowMapperFactory;
-import org.jdbi.v3.core.mapper.SingleColumnMapper;
 import org.jdbi.v3.core.mapper.reflect.internal.BeanPropertiesFactory;
+import org.jdbi.v3.core.mapper.reflect.internal.PojoMapper;
 import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties;
-import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties.PojoBuilder;
 import org.jdbi.v3.core.mapper.reflect.internal.PojoProperties.PojoProperty;
 import org.jdbi.v3.core.statement.StatementContext;
-import static org.jdbi.v3.core.mapper.reflect.ReflectionMapperUtil.anyColumnsStartWithPrefix;
-import static org.jdbi.v3.core.mapper.reflect.ReflectionMapperUtil.findColumnIndex;
-import static org.jdbi.v3.core.mapper.reflect.ReflectionMapperUtil.getColumnNames;
 
 /**
  * A row mapper which maps the columns in a statement into a JavaBean. The default
@@ -49,15 +34,8 @@ import static org.jdbi.v3.core.mapper.reflect.ReflectionMapperUtil.getColumnName
  * @deprecated this class should not be public API, use {@link org.jdbi.v3.core.statement.SqlStatement#bindBean(Object)} instead.
  */
 @Deprecated
-public class BeanMapper<T> implements RowMapper<T> {
+public class BeanMapper<T> extends PojoMapper<T> {
     static final String DEFAULT_PREFIX = "";
-
-    private static final String NO_MATCHING_COLUMNS =
-        "Mapping bean type %s didn't find any matching columns in result set";
-
-    private static final String UNMATCHED_COLUMNS_STRICT =
-        "Mapping bean type %s could not match properties for columns: %s";
-
     /**
      * Returns a mapper factory that maps to the given bean class
      *
@@ -102,120 +80,18 @@ public class BeanMapper<T> implements RowMapper<T> {
         return new BeanMapper<>(type, prefix);
     }
 
-    protected boolean strictColumnMapping; // this should be default (only?) behavior but that's a breaking change
-    protected final Class<T> type;
-    protected final String prefix;
-    private final PojoProperties<T> properties;
-    private final Map<PojoProperty<T>, BeanMapper<?>> nestedMappers = new ConcurrentHashMap<>();
+    private BeanMapper(Class<T> type, PojoProperties<T> properties, String prefix) {
+        super(type, properties, prefix);
+        strictColumnMapping = false;
+    }
 
     @SuppressWarnings("unchecked")
     BeanMapper(Class<T> type, String prefix) {
         this(type, (PojoProperties<T>) BeanPropertiesFactory.propertiesFor(type), prefix);
     }
 
-    protected BeanMapper(Class<T> type, PojoProperties<T> properties, String prefix) {
-        this.type = type;
-        this.properties = properties;
-        this.prefix = prefix.toLowerCase();
-    }
-
     @Override
-    public T map(ResultSet rs, StatementContext ctx) throws SQLException {
-        return specialize(rs, ctx).map(rs, ctx);
-    }
-
-    @Override
-    public RowMapper<T> specialize(ResultSet rs, StatementContext ctx) throws SQLException {
-        final List<String> columnNames = getColumnNames(rs);
-        final List<ColumnNameMatcher> columnNameMatchers =
-                ctx.getConfig(ReflectionMappers.class).getColumnNameMatchers();
-        final List<String> unmatchedColumns = new ArrayList<>(columnNames);
-
-        RowMapper<T> result = specialize0(ctx, columnNames, columnNameMatchers, unmatchedColumns)
-            .orElseThrow(() -> new IllegalArgumentException(String.format(NO_MATCHING_COLUMNS, type)));
-
-        if (ctx.getConfig(ReflectionMappers.class).isStrictMatching()
-            && anyColumnsStartWithPrefix(unmatchedColumns, prefix, columnNameMatchers)) {
-
-            throw new IllegalArgumentException(
-                String.format(UNMATCHED_COLUMNS_STRICT, type.getSimpleName(), unmatchedColumns));
-        }
-
-        return result;
-    }
-
-    private Optional<RowMapper<T>> specialize0(StatementContext ctx,
-                                               List<String> columnNames,
-                                               List<ColumnNameMatcher> columnNameMatchers,
-                                               List<String> unmatchedColumns) {
-        final List<RowMapper<?>> mappers = new ArrayList<>();
-        final List<PojoProperty<T>> propList = new ArrayList<>();
-
-        for (PojoProperty<T> property : properties.getProperties().values()) {
-            Nested anno = property.getAnnotation(Nested.class).orElse(null);
-
-            if (anno == null) {
-                String paramName = prefix + getName(property);
-
-                findColumnIndex(paramName, columnNames, columnNameMatchers, () -> debugName(property))
-                    .ifPresent(index -> {
-                        ColumnMapper<?> mapper = ctx.findColumnMapperFor(property.getQualifiedType())
-                            .orElseGet(() -> defaultColumnMapper(property));
-
-                        mappers.add(new SingleColumnMapper<>(mapper, index + 1));
-                        propList.add(property);
-
-                        unmatchedColumns.remove(columnNames.get(index));
-                    });
-            } else {
-                String nestedPrefix = prefix + anno.value();
-                if (anyColumnsStartWithPrefix(columnNames, nestedPrefix, columnNameMatchers)) {
-                    nestedMappers
-                        .computeIfAbsent(property, d -> new BeanMapper<>(GenericTypes.getErasedType(d.getQualifiedType().getType()), nestedPrefix))
-                        .specialize0(ctx, columnNames, columnNameMatchers, unmatchedColumns)
-                        .ifPresent(nestedMapper -> {
-                            mappers.add(nestedMapper);
-                            propList.add(property);
-                        });
-                }
-            }
-        }
-
-        if (mappers.isEmpty() && !columnNames.isEmpty()) {
-            return Optional.empty();
-        }
-
-        return Optional.of((r, c) -> {
-            final PojoBuilder<T> pojo = properties.create();
-
-            for (int i = 0; i < mappers.size(); i++) {
-                RowMapper<?> mapper = mappers.get(i);
-                PojoProperty<T> property = propList.get(i);
-
-                Object value = mapper.map(r, ctx);
-
-                pojo.set(property, value);
-            }
-
-            return pojo.build();
-        });
-    }
-
-    private ColumnMapper<?> defaultColumnMapper(PojoProperty<T> property) {
-        if (strictColumnMapping) {
-            throw new NoSuchMapperException(String.format(
-                    "Couldn't find mapper for property '%s' of type '%s' from %s", property.getName(), property.getQualifiedType(), type));
-        }
-        return (r, n, c) -> r.getObject(n);
-    }
-
-    private String getName(PojoProperty<T> property) {
-        return property.getAnnotation(ColumnName.class)
-                .map(ColumnName::value)
-                .orElseGet(property::getName);
-    }
-
-    private String debugName(PojoProperty<T> p) {
-        return String.format("%s.%s", type.getSimpleName(), p.getName());
+    protected BeanMapper<?> createNestedMapper(StatementContext ctx, PojoProperty<T> property, String nestedPrefix) {
+        return new BeanMapper<>(GenericTypes.getErasedType(property.getQualifiedType().getType()), nestedPrefix);
     }
 }
